% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{data_clean}
\alias{data_clean}
\title{Clean a dataset to make modeling fitting more efficient}
\usage{
data_clean(data, model = NULL, vars = NULL, response = "case", prev = NULL)
}
\arguments{
\item{data}{A data set that was used to fit the model}

\item{model}{A glm (with logistic or log link, with binomial family), clogit or coxph model.}

\item{vars}{Default NULL.  Variables required in output data set}

\item{prev}{Default NULL.  Prevalence of disease (or yearly incidence of disease in healthy controls).  Only relevant to set in case control studies and if path specific PAF or sequential joint PAF calculations are required.  The purpose of this is to create a vector of weights that reweights the cases and controls to reflect the general population}
}
\value{
A cleaned dataset
}
\description{
Strip out unneeded variables from original data (based on fitted model), and remove rows with NA values.  If  The function works for logistic, survival and conditional logistic regressions.  The function also creates a column of weights
}
\examples{
model <- glm(formula = case ~ regionnn7*ns(eage, df = 5) + esex*ns(eage, df = 5) + subeduc +phys + ns(aheiscore, df = 3) + nevfcur + alcohfreqwk + global_stress2 + ns(apob_apoa,df = 3) + ns(whr, df = 3) + htnadmbp, family = "binomial", data = stroke_reduced)
unclean_stroke <- rbind(stroke_reduced,rep(NA,ncol(stroke_reduced)+1))
clean_stroke <- data_clean(model, unclean_stroke)
dim(clean_stroke)
dim(unclean_stroke)
}
