% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_PAF.R
\name{do_sim}
\alias{do_sim}
\title{Internal:  Simulate a column from the post intervention distribution corresponding to eliminating a risk factor}
\usage{
do_sim(colnum, current_mat, model, SN = TRUE)
}
\arguments{
\item{colnum}{The column indicator for the variable being simulated}

\item{current_mat}{The current value of the data frame}

\item{model}{A fitted model for simulating values of the variable, given the parent values}

\item{SN}{Logical.  If TRUE (default) simulations are achieved via adding the original model residuals, to the new fitted values based on the updated values of parents in current_mat.}
}
\value{
An updated data frame (a new version of current_mat) with a new column simulated
}
\description{
Internal:  Simulate a column from the post intervention distribution corresponding to eliminating a risk factor
}
