% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanplot.R
\name{fan_plot}
\alias{fan_plot}
\title{Create a fan_plot of a rf.data.frame object}
\usage{
fan_plot(rf_data_frame)
}
\arguments{
\item{rf_data_frame}{A rf.data.frame object}
}
\value{

}
\description{
Create a fan plot displaying approximate PAF, risk factor prevalence and risk ratios
}
\examples{
rfs <- rf_summary(rf_names=c('Hypertension','Inactivity','ApoB/ApoA','Diet','WHR','Smoking','Cardiac causes','Alcohol','Global Stress','Diabetes'),rf_prev=c(.474,.837,.669,.67,.67,.224,.049,.277,.144,.129),risk=c(1.093,0.501,0.428,0.378,0.294,0.513,1.156,0.186,0.301,0.148),log=TRUE)
fan_plot(rfs)
}
