% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hordaland_data}
\alias{Hordaland_data}
\title{Simulated case control dataset for 5000 cases (individuals with chronic cough) and 5000 controls}
\format{
A data frame with 10000 rows and 4 variables:
\describe{
\item{y}{Chronic Cough, 1: Yes, 0: No}
\item{urban.rural}{1: resident in urban setting, 0: resident in rural setting}
\item{smoking.category}{Smoking level: 1 No smoker, 2: ex smoker, 3: 1-9 cigarettes per day, 4: 10-19 cigarettes per day, 4:>= 20 cigarettes per day}
\item{occupational.exposure}{Exposed to dust/gas at work.  1: Yes, 0: no}

}
}
\source{
Data simulated based on "Sequential and average attributable fractions as aids in the selection of preventive strategies." Journal of clinical epidemiology 48, no. 5 (1995): 645-655.
}
\usage{
Hordaland_data
}
\description{
Simulated case control dataset for 5000 cases (individuals with chronic cough) and 5000 controls
}
\keyword{datasets}
