% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_PSPAF.R
\name{joint_pspaf}
\alias{joint_pspaf}
\title{Calculation of joint paf taking into account risk factor sequencing}
\usage{
joint_pspaf(
  data,
  model_list,
  parent_list,
  node_vec,
  prev = 0.09,
  vars = NULL,
  response_model,
  mediator_models,
  riskfactor,
  refval,
  calculation_method,
  ci = FALSE,
  boot_rep = 100,
  ci_type = c("norm"),
  ci_level = 0.95
)
}
\arguments{
\item{data}{Data frame. A dataframe containing variables used for fitting the models.  Must contain all variables used in fitting}

\item{model_list}{List.  A list of models corresponding for the outcome variables in node_vec, with parents as described in parent_vec.  This list must be in the same order as node_vec and parent_list}

\item{parent_list}{A list.  The ith element is the vector of variable names that are direct causes of ith variable in node_vec}

\item{node_vec}{A vector corresponding to the nodes in the Bayesian network.  This must be specified from root to leaves - that is ancestors in the causal graph for a particular node are positioned before their descendants.  If this condition is false the function will return an error.}

\item{ci}{Logical. If TRUE, a bootstrap confidence interval is computed along with a point estimate (default FALSE).  If ci=FALSE, only a point estimate is produced.  A simulation procedure (sampling permutations and also simulating the effects of eliminating risk factors over the descendent nodes in a Bayesian network) is required to produce the point estimates.  The point estimate will change on repated runs of the function.  The margin of error of the point estimate is given when ci=FALSE}

\item{boot_rep}{Integer.  Number of bootstrap replications (Only necessary to specify if ci=TRUE)}

\item{ci_type}{Character.  Default norm.  A vector specifying the types of confidence interval desired.  "norm", "basic", "perc" and "bca" are the available metho}

\item{ci_level}{Numeric.  Confidence level.  Default 0.95}
}
\description{
Calculation of joint paf taking into account risk factor sequencing
}
\examples{
set.seed(15122021)
library(dplyr)
library(devtools)
library(splines)
library(survival)
library(parallel)
options(boot.parallel="snow")
options(boot.ncpus=parallel::detectCores())
parent_exercise <- c("education")
parent_diet <- c("education")
parent_smoking <- c("education")
parent_alcohol <- c("education")
parent_stress <- c("education")
parent_high_blood_pressure <- c("education","exercise","diet","smoking","alcohol","stress")
parent_lipids <- c("education","exercise","diet","smoking","alcohol","stress")
parent_waist_hip_ratio <- c("education","exercise","diet","smoking","alcohol","stress")
parent_early_stage_heart_disease <- c("education","exercise","diet","smoking","alcohol","stress","lipids","waist_hip_ratio","high_blood_pressure")
parent_diabetes <- c("education","exercise","diet","smoking","alcohol","stress","lipids","waist_hip_ratio","high_blood_pressure")
parent_case <- c("education","exercise","diet","smoking","alcohol","stress","lipids","waist_hip_ratio","high_blood_pressure","early_stage_heart_disease","diabetes")
parent_list <- list(parent_exercise,parent_diet,parent_smoking,parent_alcohol,parent_stress,parent_high_blood_pressure,parent_lipids,parent_waist_hip_ratio,parent_early_stage_heart_disease,parent_diabetes,parent_case)
node_vec=c("exercise","diet","smoking","alcohol","stress","high_blood_pressure","lipids","waist_hip_ratio","early_stage_heart_disease","diabetes","case")
model_list=automatic_fit(data=stroke_reduced, parent_list=parent_list, node_vec=node_vec, prev=.0035,common="region*ns(age,df=5)+sex*ns(age,df=5)", spline_nodes = c("waist_hip_ratio","lipids","diet"))
response_model <- glm(case ~ region*ns(age,df=5)+sex*ns(age,df=5) +  education + exercise + ns(waist_hip_ratio,df=5)+ smoking + alcohol + stress + high_blood_pressure + ns(lipids, knots = quantile(lipids,c(.25,0.5,0.75),na.rm=TRUE), Boundary.knots = quantile(lipids,c(.001,0.90),na.rm=TRUE))+ns(waist_hip_ratio,df=5)+diet,weights=weights, data=stroke_reduced,family='binomial')
mediator_models <- list(
  glm(formula = high_blood_pressure ~ region * ns(age, df = 5) + sex*ns(age, df = 5) + education + exercise + diet + smoking + alcohol + stress, family = "binomial", data = stroke_reduced, weights = weights),
  lm(formula = lipids ~ region * ns(age, df = 5) + sex * ns(age, df = 5) + education + exercise + diet + smoking + alcohol + stress, data = stroke_reduced, weights = weights),
  lm(formula = waist_hip_ratio ~ region * ns(age, df = 5) + sex * ns(age,df = 5) + education + exercise + diet + smoking + alcohol + stress,weights = weights,data=stroke_reduced))
## No confidence interval, no bootstrap
# average_pspaf_no_CI(data=stroke_reduced, model_list=model_list, parent_list=parent_list, node_vec=node_vec, prev = 0.0035/0.9965, nsim=NULL, correct_order = NULL, alpha=0.05,
                  # vars = c("exercise","high_blood_pressure","lipids","waist_hip_ratio"), exact = TRUE, response_model = response_model, mediator_models = mediator_models,
                  # riskfactor = "exercise", refval=0, calculation_method = "D")
## Bootstrap
# # average_pspaf(data=stroke_reduced, model_list=model_list, parent_list=parent_list, node_vec=node_vec, prev = 0.0035/0.9965, nsim=NULL, correct_order = 3,
                #  vars = c("exercise","high_blood_pressure","lipids","waist_hip_ratio"), exact = TRUE, response_model = response_model, mediator_models = mediator_models,
                 # riskfactor = "exercise", refval=0, calculation_method = "D", ci=TRUE,boot_rep=100, ci_type=c("norm"),ci_level=0.95, ci_level_ME=0.95)

joint_pspaf(data=stroke_reduced, model_list=model_list, parent_list=parent_list, node_vec=node_vec, prev = 0.0035/0.9965, vars = c("exercise","high_blood_pressure","lipids","waist_hip_ratio"),response_model = response_model, mediator_models = mediator_models, riskfactor = "exercise", refval=0, calculation_method = "D", ci=FALSE,boot_rep=100, ci_type=c("norm"),ci_level=0.95)
}
