% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomogram.R
\name{nomogram}
\alias{nomogram}
\title{Create a PAF nomogram of a rf.data.frame object}
\usage{
nomogram(rf_data_frame)
}
\arguments{
\item{rf_data_frame}{}
}
\value{

}
\description{
Create a PAF nomogram of a rf.data.frame object
}
\examples{

rfs <- rf_summary(rf_names=c('Hypertension','Inactivity','ApoB/ApoA','Diet','WHR','Smoking','Cardiac causes','Alcohol','Global Stress','Diabetes'),rf_prev=c(.474,.837,.669,.67,.67,.224,.049,.277,.144,.129),risk=c(1.093,0.501,0.428,0.378,0.294,0.513,1.156,0.186,0.301,0.148),log=TRUE)
nomogram(rfs)
}
