% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAF_calc_continuous.R
\name{PAF_calc_continuous}
\alias{PAF_calc_continuous}
\title{Calculation of attributable fraction with a continuous risk factor}
\usage{
PAF_calc_continuous(
  model,
  riskfactor_vec,
  q_vec = c(0.01),
  data,
  calculation_method = "B",
  prev = NULL,
  ci = FALSE,
  boot_rep = 10,
  t_vector = NULL,
  ci_level = 0.95,
  ci_type = c("norm"),
  S = 1
)
}
\arguments{
\item{model}{Either a clogit, glm or survival regression}

\item{riskfactor_vec}{A vector of continuous risk names.}

\item{q_vec}{A vector of 'risk quantiles' for the continuous exposure.  q_vec=c(0.01) (the default) calculates an estimate of the PAF, and is in some way analogous to eliminating a categorical risk factor.  Other values in q_vec correspond to interventions on the continuous risk factor that results in a risk level for all individuals thresholded above by the corresponding quantile of the pre-intervention population risk.  For survival regressions only single element values of q_vec are allowed}

\item{data}{A dataframe containing variables used for fitting the model}

\item{calculation_method}{A character either 'B' (Bruzzi) or 'D' (Direct method).  For case control data, the method described in Bruzzi 1985 is recommended.  Bruzzi's method estimates PAF from relative risks and prevalence of exposure to the risk factor.  The Direct method estimates PAF by summing estimated probabilities of disease in the absense of exposure on the individual level}

\item{prev}{The estimated prevalence of disease (A number between 0 and 1).  This only needs to be specified if the data source is from a case control study, and the direct method is used}

\item{ci}{Logical. If TRUE, a bootstrap confidence interval is computed along with point estimate (default FALSE)}

\item{t_vector}{Numeric.  A vector of times at which to calculate PAF (only specified if model is coxph)}

\item{ci_level}{Numeric.  A number between 0 and 1 specifying the confidence level}

\item{ci_type}{Character.  A vector specifying the types of confidence interval desired, as available in the 'Boot' package. The default is c('norm'), which calculates a symmetric confidence interval: (Est-Bias +- 1.96*SE), with the standard error calculated via Bootstrap.  Other choices are 'basic', 'perc' and 'bca'.  Increasing the number of Bootstrap repetitions is recommended for the 'basic', 'perc' and 'bca' methods.}

\item{S}{Integer.  The number of randomly selected individuals for which risk is measured (defaults to 1).  Let to perhaps 100 if risk factor involved in interactions in model}

\item{nboot}{Integer.  Number of bootstrap replications (Only necessary to specify if ci=TRUE)}
}
\value{
A vector of estimated PAF if ci=FALSE, corresponding to the quantiles specified in q_vec.  If ci=TRUE, a data frame with columns corresponding to the raw estimate, estiamted bias, bias corrected estimate and lower and upper elements of any confidence procedures requested, and rows corresopnding to the quantiles in q_vec.
}
\description{
Calculation of attributable fraction with a continuous risk factor
}
\examples{
library(splines)
library(survival)
library(parallel)
options(boot.parallel="snow")
options(boot.ncpus=parallel::detectCores())
# Example with logistic regression.  PAF_q (as in Ferguson, 2020) estimated at q=0.01, 0.1, 0.3, 0.5, 0.7. 0.9.  PAF_0.01 is roughly analogous to 'eliminating' a discrete risk factor, but its estimation may be unstable for some exposures, and the corresponding intervention may be impractical.  Comparing PAF_q for q >= 0.1 over different risk factors may lead to more sensible comparisons of disease burden.  Either method (direct, D, or Bruzzi )
model_continuous <- glm(formula = case ~ region * ns(age, df = 5) + sex * ns(age, df = 5) + education +exercise + ns(diet, df = 3) + alcohol + stress + ns(lipids,df = 3) + ns(waist_hip_ratio, df = 3) + high_blood_pressure, family = "binomial", data = stroke_reduced)

out <- PAF_calc_continuous(model_continuous,riskfactor_vec=c("diet","lipids","waist_hip_ratio"),q_vec=c(0.01, 0.1,0.3,0.5,0.7,0.9),ci=TRUE,calculation_method="D",data=stroke_reduced, prev=0.0035)

 # The same example but using conditional logsitic regression.  B
model_continuous_clogit <- clogit(formula = case ~ region * ns(age, df = 5) + sex * ns(age, df = 5) + education +exercise + ns(diet, df = 3)  + alcohol + stress + ns(lipids,df = 3) + ns(waist_hip_ratio, df = 3) + high_blood_pressure + strata(strata), data = stroke_reduced)
PAF_calc_continuous(model_continuous_clogit,riskfactor_vec=c("diet","lipids","waist_hip_ratio"),q_vec=c(0.01, 0.1,0.3,0.5,0.7,0.9),ci=TRUE,calculation_method="B",data=stroke_reduced, prev=0.01)
}
