% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAF_calc_discrete.R
\name{PAF_calc_discrete}
\alias{PAF_calc_discrete}
\title{Calculation of attributable fraction using a categorized risk factor}
\usage{
PAF_calc_discrete(
  model,
  riskfactor,
  refval,
  data,
  calculation_method = "B",
  prev = NULL,
  ci = FALSE,
  boot_rep = 100,
  t_vector = NULL,
  ci_level = 0.95,
  ci_type = c("norm"),
  ind
)
}
\arguments{
\item{model}{Either a clogit, glm or coxph survival regression}

\item{riskfactor}{The name of the risk factor of interest in the dataset.  The risk factor values can be 0/1 numeric, categorical or factor valued.}

\item{refval}{The reference value for the risk factor.  If a risk factor is 0/1 numeric, 0 is assumed as the default value, otherwise refval must be specified.}

\item{data}{A dataframe containing variables used for fitting the model}

\item{calculation_method}{A character either 'B' (Bruzzi) or 'D' (Direct method).  For case control data, the method described in Bruzzi 1985 is recommended.  Bruzzi's method estimates PAF from relative risks and prevalence of exposure to the risk factor.  The Direct method estimates PAF by summing estimated probabilities of disease in the absense of exposure on the individual level}

\item{prev}{The estimated prevalence of disease (A number between 0 and 1).  This only needs to be specified if the data source is from a case control study, and the direct method is used}

\item{ci}{Logical. If TRUE, a bootstrap confidence interval is computed along with point estimate (default FALSE)}

\item{boot_rep}{Integer.  Number of bootstrap replications (Only necessary to specify if ci=TRUE)}

\item{t_vector}{Numeric.  A vector of times at which to calculate PAF (only specified if model is coxph)}

\item{ci_level}{Numeric.  A number between 0 and 1 specifying the confidence level}

\item{ci_type}{Character.  A vector specifying the types of confidence interval desired, as available in the 'Boot' package. The default is c('norm'), which calculates a symmetric confidence interval: (Est-Bias +- 1.96*SE), with the standard error calculated via Bootstrap.  Other choices are 'basic', 'perc' and 'bca'.  Increasing the number of Bootstrap repetitions is recommended for the 'basic', 'perc' and 'bca' methods.}
}
\value{
An estimated PAF if ci=FALSE, or for survival data a vector of estimated PAF corresponding to event times in the data.  If ci=TRUE, a vector with elements corresponding to the raw estimate, estiamted bias, bias corrected estimate and lower and upper elements of any confidence procedures requested.  If ci=TRUE, and a coxph model is fit, a matrix will be returned, with rows corresponding to differing times at which the PAF might be calculated.
}
\description{
Calculation of attributable fraction using a categorized risk factor
}
\examples{
data(stroke_reduced)
library(splines)
model_exercise <- glm(formula = case ~ region * ns(age, df = 5) + sex * ns(age, df = 5) + education + exercise + ns(diet, df = 3) + smoking + alcohol + stress, family = "binomial", data = stroke_reduced)
# calculate discrete PAF using Bruzzi method
PAF_calc_discrete(model_exercise, "exercise", refval=0, data=stroke_reduced, calculation_method="B")
# calculate discrete PAF using Direct method
# PAF_calc_discrete(model_exercise, "exercise", refval=0, data=stroke_reduced, calculation_method="D", prev=0.0035)
# Use bootstrap resampling to calculate a confidence interval
PAF_calc_discrete(model_exercise, "exercise", refval=0, data=stroke_reduced, calculation_method="D", prev=0.005, ci=TRUE, boot_rep=10)
### use the Bruzzi method derived by Bruzzi, 1985, instead
PAF_calc_discrete(model_exercise, "exercise", refval=0, data=stroke_reduced, calculation_method="B", ci=TRUE, boot_rep=10)
}
