% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levin.R
\name{paf_levin}
\alias{paf_levin}
\title{Levin's formula based on relative risk and prevalence}
\usage{
paf_levin(prev = NULL, RR = NULL, conf_prev = NULL, conf_RR = NULL, digits = 3)
}
\arguments{
\item{prev}{Estimated prevalence.  Can be left unspecified if conf_prev specified.}

\item{RR}{Estimated relative risk.  Can be left unspecified if conf_RR specified.}

\item{conf_prev}{A numeric vector of length 2 giving confidence limits for prevalence.}

\item{conf_RR}{A numeric vector of length 2 giving the confidence limits for relative risk.#'}

\item{digits}{integer.  The number of significant digits for rounding of PAF estimates and confidence intervals.  Default of 3.}
}
\value{
If confidence intervals for prevalence and relative risk are not specified, the estimated PAF.  If confidence intervals for prevalence and relative risk are specified, confidence intervals for PAF are estimated using approximate propagation of imprecision.  Note that if confidence intervals are supplied as arguments, the algorithm makes assumptions that the point estimate of prevalence is the average of the specified confidence limits for prevalence, the point estimate for relative risk is the geometric mean of the confidence limits for relative risk, and that the 3 estimators are independent.
}
\description{
Levin's formula based on relative risk and prevalence
}
\examples{
CI_p <- c(0.1,0.3)
CI_RR <- c(1.2, 2)
# calculation without confidence interval
paf_levin(prev=0.2,RR=exp(.5*log(1.2)+.5*log(2)))
# calculation with confidence interval
paf_levin(conf_prev=CI_p,conf_RR=CI_RR)
}
