% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levin.R
\name{paf_miettinen}
\alias{paf_miettinen}
\title{Miettinen's formula based on adjusted relative risk, unadjusted relative risk and prevalence}
\usage{
paf_miettinen(
  prev = NULL,
  RR = NULL,
  RR_u = NULL,
  conf_prev = NULL,
  conf_RR = NULL,
  conf_RRu = NULL,
  digits = 3
)
}
\arguments{
\item{prev}{Estimated prevalence.  Can be left unspecified if conf_prev specified.}

\item{RR}{Estimated adjusted relative risk.  Can be left unspecified if conf_RR specified.}

\item{RR_u}{Estimated unadjusted relative risk.  Can be left unspecified if conf_RRu specified.}

\item{conf_prev}{A numeric vector of length 2 giving confidence limits for prevalence.}

\item{conf_RR}{A numeric vector of length 2 giving the confidence limits for the adjusted relative risk.}

\item{conf_RRu}{A numeric vector of length 2 giving the confidence limits for the unadjusted relative risk.}

\item{digits}{integer.  The number of significant digits for rounding of PAF estimates and confidence intervals.  Default of 3.}
}
\value{
If confidence intervals for prevalence and adjusted and unadjusted relative risk are not specified, the estimated PAF.  If confidence intervals are specified, confidence intervals for PAF are also estimated using approximate propagation of imprecision.  Note that if confidence intervals are supplied as arguments, the algorithm makes assumptions that the point estimate of prevalence is the average of the specified confidence limits for prevalence, the point estimates for adjusted/unadjusted relative risk are the geometric means of the specified confidence limits for relative risk, and that the 3 estimators are independent.
}
\description{
Miettinen's formula based on adjusted relative risk, unadjusted relative risk and prevalence
}
\examples{
CI_p <- c(0.1,0.3)
CI_RR <- c(1.2, 2)
CI_RRu <- c(1.5, 2.5)
# example without confidence interval
paf_miettinen(prev=0.2,RR=exp(.5*log(1.2)+.5*log(2)), RR_u=exp(.5*log(1.5)+.5*log(2.5)))
#' # example with confidence interval
paf_miettinen(conf_prev=CI_p,conf_RR=CI_RR, conf_RRu=CI_RRu)
}
