% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PAF_continuous.R
\name{plot_PAF_q}
\alias{plot_PAF_q}
\title{Plot of impact fractions against over risk-quantile interventions for several risk factors}
\usage{
plot_PAF_q(data_frame)
}
\arguments{
\item{data_frame}{A data frame object, the final 3 columns of which are assumed to be (in order), PAF and lower and upper confidence bounds.}
}
\value{

}
\description{
Plot of impact fractions against over risk-quantile interventions for several risk factors
}
\examples{


}
