% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAF_calc_continuous.R
\name{predict_df_continuous}
\alias{predict_df_continuous}
\title{Internal:  Create a data frame for predictions (when risk factor is continuous).}
\usage{
predict_df_continuous(riskfactor, q_val, risk_q, data)
}
\arguments{
\item{riskfactor}{The name of the risk factor of interest in the dataset}

\item{q_val}{The risk quantile to match to}

\item{risk_q}{Estimated risk quantiles}

\item{data}{A dataframe containing variables used to fit the model}
}
\value{
A data frame where the distribution continuous risk factor so at an individual level, risk is at the q_val-quantile or below
}
\description{
Internal:  Create a data frame for predictions (when risk factor is continuous).
}
