% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAF_calc_discrete.R
\name{predict_df_discrete}
\alias{predict_df_discrete}
\title{Internal:  Create a data frame for predictions (when risk factor is discrete).}
\usage{
predict_df_discrete(riskfactor, refval, data)
}
\arguments{
\item{riskfactor}{The name of the risk factor of interest in the dataset}

\item{refval}{The reference value for the risk factor}

\item{data}{A dataframe containing variables used to fit the model}
}
\value{
A data frame where the categorical variable is set to its reference level
}
\description{
Internal:  Create a data frame for predictions (when risk factor is discrete).
}
