% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_PAF.R
\name{refit}
\alias{refit}
\title{Internal:  refit a model based on new dataframe}
\usage{
refit(model, data, with_weights = FALSE)
}
\arguments{
\item{model}{Original model used (either lm, glm, polr, coxph or clogit)}

\item{data}{New data on which to refit the same model}

\item{with_weights}{Logical.  If TRUE fit uses weighted maximum lieklihood, based on a column 'weights' that is assumed to be specified in the data frame}
}
\value{
An updated fitted model, based on the new dataframe
}
\description{
Internal:  refit a model based on new dataframe
}
