% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanplot.R
\name{rf_summary}
\alias{rf_summary}
\title{Create a summary data frame for risk factors}
\usage{
rf_summary(rf_names, rf_prev, risk, log = TRUE)
}
\arguments{
\item{rf_names}{character}

\item{rf_prev}{double}

\item{risk}{double}

\item{log}{logical}
}
\value{
A rf.data.frame object
}
\description{
Create a summary data frame for risk factors, prevalence and risk ratios.  This will be used in fan plots and nomograms
}
\examples{

rfs <- rf_summary(rf_names=c('Hypertension','Inactivity','ApoB/ApoA','Diet','WHR','Smoking','Cardiac causes','Alcohol','Global Stress','Diabetes'),rf_prev=c(.474,.837,.669,.67,.67,.224,.049,.277,.144,.129),risk=c(1.093,0.501,0.428,0.378,0.294,0.513,1.156,0.186,0.301,0.148),log=TRUE)
}
