% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PAF_continuous.R
\name{plot_PAF_q}
\alias{plot_PAF_q}
\title{Plot of impact fractions against over risk-quantile interventions for several risk factors}
\usage{
plot_PAF_q(data_frame)
}
\arguments{
\item{data_frame}{A data frame object, the final 3 columns of which are assumed to be (in order), PAF and lower and upper confidence bounds.}
}
\value{

}
\description{
Plot of impact fractions against over risk-quantile interventions for several risk factors
}
\examples{
library(splines)
library(survival)
library(parallel)
options(boot.parallel="snow")
options(boot.ncpus=parallel::detectCores())
model_continuous <- glm(formula = case ~ region * ns(age, df = 5) + sex * ns(age, df = 5) + education +exercise + ns(diet, df = 3) + alcohol + stress + ns(lipids,df = 3) + ns(waist_hip_ratio, df = 3) + high_blood_pressure, family = "binomial", data = stroke_reduced)
out <- PAF_calc_continuous(model_continuous,riskfactor_vec=c("diet","lipids","waist_hip_ratio"),q_vec=c(0.01, 0.1,0.3,0.5,0.7,0.9),ci=TRUE,calculation_method="B",data=stroke_reduced)
plot_PAF_q(out)
}
