% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspaf.R
\name{ps_paf}
\alias{ps_paf}
\title{General calculations of impact fractions}
\usage{
ps_paf(
  response_model,
  mediator_models,
  riskfactor,
  refval,
  data,
  prev = NULL,
  ci = FALSE,
  boot_rep = 100,
  ci_level = 0.95,
  ci_type = c("norm")
)
}
\arguments{
\item{response_model}{A R model object for a binary outcome that involves a risk factor, confounders and mediators of the risk factor outcome relationship}

\item{mediator_models}{A list of R object models for the mediator relationship (depending on the risk factor and any confounders)}

\item{riskfactor}{character.  Represents the name of the risk factor}

\item{refval}{For factor valued risk factors, the refernece level of the risk factor.  If the risk factor is numeric, the reference level is assumed to be 0.}

\item{data}{dataframe. A dataframe (with no missing values) containing the data used to fit the mediator and response models.  Run data_clean on dataset before using}

\item{prev}{numeric.  A value between 0 and 1 specifying the prevalence of disease.}

\item{ci}{logical.  If TRUE a confidence interval is calculated using Bootstrap}

\item{boot_rep}{Integer.  Number of bootstrap replications (Only necessary to specify if ci=TRUE)}

\item{ci_level}{Numeric.  Default 0.95. A number between 0 and 1 specifying the confidence level (only necessary to specify when ci=TRUE)}

\item{ci_type}{Character.  Defalt norm.  A vector specifying the types of confidence interval desired.  "norm", "basic", "perc" and "bca" are the available methods}

\item{calculation_method}{A character either 'B' (Bruzzi) or 'D' (Direct method).  For case control data, the method described in Bruzzi 1985 is recommended.  Bruzzi's method estimates PAF from relative risks and prevalence of exposure to the risk factor.  The Direct method estimates PAF by summing estimated probabilities of disease in the absense of exposure on the individual level}
}
\value{
An estimated impact if ci=FALSE, or for survival data a vector of estimated impact corresponding to event times in the data.  If ci=TRUE, a vector with elements corresponding to the raw estimated impact fraction, estiamted bias, bias corrected estimate and lower and upper elements of any confidence procedures requested.  If ci=TRUE, and a coxph model is fit, a matrix will be returned, with rows corresponding to differing times at which the impact fraction might be calculated.
}
\description{
General calculations of impact fractions
}
\examples{
library(splines)
library(survival)
library(parallel)
options(boot.parallel="snow")
options(boot.ncpus=parallel::detectCores())
# Direct and pathway specific attributable fractions estimated on simulated case control stroke data:
# Note that the nodes here are weighted regressions (based on a column in the dataframe named 'weights') to rebalance the case control structure to make it representative over the population, according to the prev argument.
# Unweighted regression is fine to use if the data arises from cohort or cross sectional studies, in which case prev should be set to NULL
response_model <- glm(case ~ region * ns(age, df = 5) + sex * ns(age, df = 5) + education + exercise + ns(diet, df = 3) + smoking + alcohol + stress + ns(lipids, df = 3) + ns(waist_hip_ratio, df = 3) + high_blood_pressure, data=stroke_reduced,family='binomial', weights=weights)
mediator_models <- list(glm(high_blood_pressure ~ region * ns(age, df = 5) + sex * ns(age, df = 5) + education   +exercise + ns(diet, df = 3) + smoking + alcohol + stress,data=stroke_reduced,family='binomial',weights=weights),lm(lipids ~ region * ns(age, df = 5) + sex * ns(age, df = 5) +education +  exercise + ns(diet, df = 3) + smoking + alcohol + stress, weights=weights, data=stroke_reduced),lm(waist_hip_ratio ~ region * ns(age, df = 5) + sex * ns(age, df = 5) + education + exercise + ns(diet, df = 3) + smoking + alcohol + stress, weights=weights, data=stroke_reduced))
ps_paf(response_model=response_model, mediator_models=mediator_models ,riskfactor="exercise",refval=0,data=stroke_reduced,prev=0.0035, ci=TRUE,boot_rep=100,ci_type="norm")
}
